Implementando la parte de solicitad calificaciones y las acciones de la ultima accion

+++ ProyectoEstructura/src/main/java/POJOs/SolicitudCalificacion.java
@@ -34,7 +34,7 @@ public class SolicitudCalificacion {
             
         }
         // Agrega la calificacion a la lista del estudiante
-        estudiante.getCalificaciones().add((int)calificacion) ;
+        estudiante.getCalificaciones().add(calificacion) ;
         // se Crea un objeto de tipo Accion
         Accion accionCalificar = new Accion("CALIFICACION", this.estudiante);
         // Envia la accion a la pila
